requirements.txt
playwright
pytest
pytest-bdd
allure-pytest
python-dotenv
pytest-playwright
pytest-dotenv

